JobCard child component created and added to parent BrowseJobs
